Extract function-call scope effects into FuncCallScopeEffectsHelper
The scope-only side effects of a call - by-ref array function results
(array_pop(), sort(), array_splice(), ...), paired-read invalidation
(json_last_error(), file_get_contents()), possibly-impure value remembering,
output-buffer level tracking and volatile-expression invalidation - move out
of FuncCallHandler::processExpr() into a dedicated helper. Nothing in the
moved chain touches the call's own result state.

processExpr() runs for every function call and its frame is paid on every
recursion level; the extraction shrinks it from 1,536 to 960 bytes
(91 -> 55 compiled variables) and FuncCallHandler from 1,338 to ~900 lines.
